viddycast, a hls live stream powered alternative
viddycast is a variant which uses hls live streaming (instead of hls
video on demand). the video is much less synced between different
users, but it should be more compatible, and only takes up a few
megabytes on the server during playback, with no need to encode ahead
of the event. It’s well suited for simply playing a video file from
your server across an unreliable network. Nothing interesting, but it
sure is easy to use compared to wrangling ffmpeg commands yourself